chore(swap): cleanup swap controller and remove usused ABIs
