+++ app/scripts/metamask-controller.js
@@ -420,11 +420,7 @@
        getNextNonce: nodeify(this.getNextNonce, this),
  
        // swapsController
-       wrapETH: nodeify(swapsController.wrapETH, swapsController),
-       // quote: nodeify(swapsController.quote, swapsController),
        quote: nodeify(this.quote, this),
-       approveTokenAllowance: nodeify(swapsController.approveTokenAllowance, swapsController),
-       fillOrder: nodeify(this.fillOrder, this),
  
        // messageManager
        signMessage: nodeify(this.signMessage, this),
